refactor: migrate test ipc-contract imports to ipc-protocol
