+++ assistant/src/__tests__/handlers-cu-observation-blob.test.ts
@@ -54,7 +54,7 @@ import type {
   CuObservation,
   IpcBlobRef,
   ServerMessage,
-} from "../daemon/ipc-contract.js";
+} from "../daemon/ipc-protocol.js";
 import { DebouncerMap } from "../util/debounce.js";
 
 /** Poll until a predicate is true or timeout (default 2s). */
