+++ assistant/src/__tests__/notification-deep-link.test.ts
@@ -67,7 +67,7 @@ mock.module("../notifications/conversation-pairing.js", () => ({
   },
 }));
 
-import type { ServerMessage } from "../daemon/ipc-contract.js";
+import type { ServerMessage } from "../daemon/ipc-protocol.js";
 import { VellumAdapter } from "../notifications/adapters/macos.js";
 import { NotificationBroadcaster } from "../notifications/broadcaster.js";
 import type { NotificationSignal } from "../notifications/signal.js";
