+++ assistant/src/__tests__/oauth-connect-handler.test.ts
@@ -119,7 +119,7 @@ import { handleMessage, type HandlerContext } from "../daemon/handlers.js";
 import type {
   OAuthConnectStartRequest,
   ServerMessage,
-} from "../daemon/ipc-contract.js";
+} from "../daemon/ipc-protocol.js";
 import { DebouncerMap } from "../util/debounce.js";
 
 function createTestContext(): { ctx: HandlerContext; sent: ServerMessage[] } {
