+++ assistant/src/__tests__/twitter-auth-handler.test.ts
@@ -177,7 +177,7 @@ import type {
   ServerMessage,
   TwitterAuthStartRequest,
   TwitterAuthStatusRequest,
-} from "../daemon/ipc-contract.js";
+} from "../daemon/ipc-protocol.js";
 import { DebouncerMap } from "../util/debounce.js";
 
 function createTestContext(): { ctx: HandlerContext; sent: ServerMessage[] } {
